requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
Pillow
requests
Flask
scikit-learn
nltk
scipy
matplotlib
seaborn